Update README.md
Document the changes that were made in the pull request:
https://github.com/collectiveidea/delayed_job/pull/168/files

Looking at the code in that PR I figured that the max_attempts needs to be set on the payload and I tried setting it on the custom job and it worked as expected. This feature is indeed very useful and hopefully this is well-documented and maintained in future releases. Greatly appreciate this gem!

A question - would this work with the object.delay.some_method usage. Didn't seem like it would since the DB table does not have the max_attempts as a column.