Cleanup rubocop violations
Some important notes for this:

BlockLength cop was disabled completely because of a large amount of
violations in the specs.

DuplicatedGem cop was disabled completely because we are using
conditionals to determine Gem versions and it was being triggered. I
don't see any other options for this cop so it looks like the only
option is to ignore.

YAMLLoad wants to use safe_load over load.  Not sure that's supported on
older versions

A few cops were disabled in specific code locations because I did not
know enough about the code to update and be sure that there was no risk.