Make DJ Concurrent
We recently updated delayed_job to allow workers to pickup multiple
jobs at a time. This allows workers to work those jobs off concurrently.

We have deliberately chosen a simple implementation over a robust one.
We can definitely optimize this implementation so that it does not spawn
new threads (and rather re-uses them). However, we'd like to learn about
the edges of this implementation first in production.