Add 'queue' support to Monitor
This updates `Delayed::Monitor` to `GROUP BY` both on priority and on
queue name. It will also add a `WHERE` clause if the `QUEUE`/`QUEUES`
ENV vars are specified (accessed via `Delayed::Worker.queues`).

This means that we can choose to run one monitor per `rake jobs:work`
and have those queue configs be 1:1 with the metrics that are reported.

**Caveats:**
- You'll over-emit metrics if you allow 2 different monitors to watch
  overlapping queues.
- Queues that aren't included in `Worker.queues` (or, alternatively,
  aren't the `Worker.default_queue_name`) won't get empty (`0`) metrics
  emitted. This could cause some instrumentation backends to assume that
  you have some jobs in that queue indefinitely.
  - This only applies if you _don't_ specify any queue configuration in
    the ENV. If you specify queues, you'll get all of the expected `0`
    metrics when there are no matching jobs in the table.
  - It would be possible to track "known queues" and emit metrics to
    zero them out when they effectively disappear from query results,
    but that was beyond the scope of this PR.