Move `delayed/backend/active_record.rb` to `delayed/job.rb`
This also includes some spec consolidation (the "shared example" just
becomes the primary spec for delayed/job, and I folded the other
ActiveRecord backend spec into it).

I'm leaving the `Delayed::Backend::ActiveRecord::Job` constant for now,
to preserve git history on the file across the move. We can compact it
down as a final step.