Add back the `:execute` hook to avoid breaking ddtrace plugin, plus some configuration-moving
It seems like we're pretty married to DJ lifecycle hooks for now, unless
there comes a point when we decide it's worth it to reach out and make a
bunch of changes to other gems in the ecosystem. So this adds `:execute`
back but in the `Runnable` concern instead of in the worker, because it
needs to go around the `loop`.

I did make one change, which is to pass a `nil` argument into
`:execute`, since neither semantic_logger nor ddtrace actually make use
of the worker argument in any meaningful way, and this allows the
callback to work with the `Monitor` process. I also added test coverage
for this callback.

Other changes in this PR:
- Move `Delayed::Worker.logger` to `Delayed.logger` (with delegation for
  backwards compatibility)
- Move `Delayed::Worker.default_log_level` to
  `Delayed.default_log_level` (with delegation for backwards
  compatibility)
- Move `Delayed::Worker.lifecycle` to `Delayed.lifecycle` (with
  delegation for backwards compatibility)
- Move `Delayed::Worker.plugins` to `Delayed.plugins` (with delegation
  for backwards compatibility)

^^ These feel like the right home for these, but I'm happy to debate.
There are other `cattr_accessor`s on `Delayed::Worker` that act more as
defaults for worker-specific behaviors, but these four feel like
gem-wide global configurations. Moving them allowed me to avoid having
`Monitor` call into `Worker`.