Prevent adapter patches from hiding job priority from other queue adapters
I noticed this when running tests that use the `:test` queue adapter,
but it would be true of any other ActiveJob adapters.

ActiveJob tries to call `.to_i` on priority (which doesn't work on
symbols), which is why I had originally passed it through
`provider_attributes` instead, but I realized that it's better to just
type-cast to a `Delayed::Priority` explicitly so that `to_i` works as
expected.