Change default priority to 10, max_claims to 5, and simplify worker state
After chatting offline, we decided to ditch the idea that worker
instances could hold their own config state, and instead rely on global
state for all of these configs. Things behave mostly the way they did
before now, except that worker instances _don't have setters_ for any of
these configs. The only way to change a config is to set it globally on
the class.