Priority based alerts and autoscaling
This is my attempt to incorporate the scaling metric
(`delayed.job.alert_age_percent`) into the monitor process, as well as
more general priority-based alert thresholds that can be used to
determine whether or not to fire off alerts when subscribing to the
ActiveSupport notifications.

Alert thresholds can be specified for `:age` (time in queue),
`:runtime`, and `:attempts` and are specific to each priority name. (I
considered adding global defaults too! But I wanted to see how this
played out before making it more complicated.)

Co-authored-by: Harris Effron <harris.effron@gmail.com>