Fix 'queue(s)' opts when running via rake tasks
The way these ENV var handlers were set up was overriding the default
value we set in our initializer (which runs as part of `:environment`).
This must be why we relied on a monkey patch (`def self.queues`) rather
than setting the queues config directly. Glad to finally have gotten to
the bottom of this 😌