Work around Rails' disabling of eager_load during rake tasks
Rails always sets `eager_load` to `false` inside of rake tasks, which we
never noticed because either:

- we were running the `:classic` autoloader with the single-threaded
  implementation, so autoloading worked, or
- we were running the `:zeitwerk` autoloader, which is thread-safe

On Rails 6.1, we'll want to set `rake_eager_load` to `true` in CI/prod.

Also, this includes a very, very, minor change that impacts ActiveJob's
built-in logging. They resolve to the constant name and remove the word
"Adapter" rather than the underlying adapter name. So the logs looked
like this:

```
[ActiveJob] Enqueued JobClass (Job ID: SOME-UUID) to ActiveJob(default)
```

And now they look like this:

```
[ActiveJob] Enqueued JobClass (Job ID: SOME-UUID) to Delayed(default)
```

To demonstrate the underlying reason:

```
class B; end
A = B
A.name # => 'B'

A = Class.new(B) # A now inherits from B
A.name # => 'A'
```