Ensure that events with matching payloads are only emitted ONCE with a single value
I baked this check into the matcher, so if it detects events with the
same payload but _different_ values, the tests will fail. I then used
this to fix the bug and get the tests green.

This should fix the weird issue where gauge metrics cycle between real
values and zero with no rhyme or reason -- it's because we were emitting
duplicate metrics with value zero!