Fix metrics that depend on db_time_now
🤦 🤦 🤦

I've been passively trying to figure out why a few of the metrics just
didn't seem to line up. Suddenly it struck me while I was making tea
before bed, and so I felt compelled to get a fix up.

...it's because we cached `@as_of = Job.db_time_now` when initializing
the monitor instance. I had been trying to be clever by providing a
single frame of reference for all of the metrics in a batch, but it was
_too clever_ and, silly me, I forgot that I made the monitor process
long-lived, so the frame of reference ends up getting cached until the
next deploy.

🤕  So this removes the cleverness and goes back to the old strategy of
using an ever-changing `Time.zone.now` (provided by `Job.db_time_now` -
to be consistent with the rest of the codebase) throughout the queries.