Add test for queue_with_priority
We were investigating a bug where jobs were not being enqueued with the
correct priority. To rule out `ActiveJob`'s `queue_with_priority`
config, I added this spec. It turns out that it works (the actual bug
was with `ActiveJob`'s `retry_on` behavior), but is helpful test
coverage to have!