Optimize JobWrapper queueing ActiveJobs
From one angle it's an optimization, from another angle it's a bugfix (:

We don't need to call ActiveJob::Base.deserialize during queueing the job, because we already have an instance of the class we can send methods to for finding out the #max_attempts or #queue_name. Teach JobWrapper how to take a job instance in and serialize it into job_data in that case, whilst still supporting being loaded from the database as a hash and calling deserialize lazily (existing flow).

The potential bugfix is someone implementing a second-order effect in #deserialize like using it as a hook to setup ActiveSupport::CurrentAttributes for the job from persisted data. If we invoke #deserialize during enqueuing this can change CurrentAttributes values in the web request that's queueing the job.