Add a config for enforcing a minimum job reserve interval (default: 0 for now) (#48)
This is related to #41 and #42 insofar as a kind of DB-bound "spinloop"
is still possible if a worker picks up jobs that take so little time
that the worker immediately turns around and asks for more. As of now,
there has been no way to tune the amount of time a worker should wait in
between _successful_ iterations of its run loop.

This introduces a configuration (`min_reserve_interval`) specifying a
minimum number of seconds (default: 0 as this is not a major release)
that a worker should wait in between _successful_ job reserve queries.
An existing config (`sleep_delay`) is still used to define the number of
seconds (default: 5) that a worker should wait in between _unsuccessful_
job reserve attempts (i.e. the queue is empty).

The job execution time is subtracted from `min_reserve_interval` when
the worker sleeps, and if jobs take more than `min_reserve_interval` to
complete than the worker will not sleep before the next reserve query.

/no-platform