feat: add `delayed_jobs.name` column (#59)
This adds a `delayed_jobs.name` column that will be populated by default
when each job is enqueued. While this column starts as nullable (to
ensure zero downtime is required to make this change), it will be
promoted in a future release to an expected column. **Update: this
future release will be TBD, and I've removed commitments in code to
remove it by any specific version.**

The purpose of this column is twofold:
1. Improve downstream developers' ability to manually operate the queue
without needing to query deep into the `handler` column.
2. Power richer observability & instrumentation tooling (e.g. top N job
types in the table, etc)

This is the first real schema change we've made, and as such I had to
rewire a little bit of plumbing to make it easier to drop a new
migration into an existing app (see `README`).