test: Add "golden" tests for worker/monitor SQL (#61)
This is the first in a series of PRs where I will aim to rework &
optimize this gem's query logic. I ultimately aim to add new index(es),
and generally improve the indexability & runtime performance of these
queries, and will produce a new major release at that point.

👉 But step one is to **test the SQL as-is** so that subsequent PRs can
confidently prove whether or not they produce different SQL! So that
test harness is what this PR introduces.

I tried to make it as adapter-agnostic as possible. We lose some
exactness in the case of a given adapter, but it means that I can
side-step the issue of which adapters we care to support. Ultimately,
the ORM is generating something with the same structure for each, and
_that's_ what I'm primarily testing. (Not the adapter-specific pickup
logic.)

I'll note that this "golden testing" strategy means that the `Monitor`
coverage in particular does feel very verbose, but I argue that if the
library is capable of producing N distinct queries, we should have N
tests, and look for ways to reduce the number of queries (rather than
making the tests more clever).