refactor: Add 'lock_timeout' & clarify that 'max_run_time' is a process-wide config (#62)
Ideally `max_run_time` is the same across an entire worker pool, because
otherwise workers might pick up jobs that are still held by other
workers in the pool.

So it has never made sense (based on our current pickup/monitor logic)
that this would be configurable on a per-worker-instance or per-query
basis.