test: Snapshot test all EXPLAIN query plans (#69)
This adds adapter-specific EXPLAIN outputs to the snapshot files introduced in #67.

The goal is to test that available indexes are usable, and enable us to flag changes/regressions in query plans going forward. **The goal is NOT to test or predict performance.** These are essentially "unit" tests for our SQL -- any integration testing must be done with a realistic dataset at a reasonable scale (outside the scope of this change).

Most of the magic happens in the adapter-specific `_explain` methods, where I use a few tricks to convince the query planners to produce stable outputs.
- **postgres**:
    - a high `seq_page_cost` value means it will favor indexes.
    - turning `enable_hashagg` off means it will never use internal stats to pick a `HashAggregate` as an optimization over an explicit `Sort` -> `GroupAggregate`.
    - `plan_cache_mode= force_generic_plan` means it will not look at parameter values to determine whether to recompute a query plan if it already has one cached.
- **mysql**:
    - Setting `max_seeks_for_key` to a low value means it will favor indexes, but it does need a sampling of rows in place for it to determine that `max_seeks` will be `> 1` without an index.
- **sqlite**: The query plans appeared stable to me, so I wasn't finding a need to attempt to trick it into doing anything different.

This depends on #68 being merged first (mostly because I want to use mysql's new "TREE" plan formatter).