fix(monitor): exclude failed jobs from metrics that shouldn't count them (#73)
This fixes a longstanding bug in the monitor queries that caused `future_count`, `locked_count`, and `erroring_count` to include failed jobs.

(The latter sort of comes down to semantics, but "erroring" to me implies that it has not yet terminally failed. I also checked and it was out of alignment with the definition we use for our internal dashboard -- on our dashboard, `erroring` also excludes failed jobs.)

While typically failed jobs don't have a `locked_at` set or a `run_at` in the future, it can sometimes happen as a result of manual operations, and has caused dashboard charts to get stuck at a nonzero baseline.

💪 Conveniently, this also improves the query plans for all three fields. (`future_count` even becomes an `Index Only` (postgres) / `Covering index` (mysql) scan/lookup!