perf(monitor): use 'attempts > 0' for better indexability (#76)
Postgres and Mysql both indicate that this enables them to use an index-only/covering scan for the `erroring` scope/metric 👍 

It would be perfectly possible to index on `last_error`, but indexing an `integer` column is a lot more straightforward than the strategies we'd need to employ for a `text` index to not be unnecessarily expensive from a writes/storage standpoint. (E.g. we could index on `WHERE last_error IS NOT NULL` for DBs that support indexes on expressions, but we could also do that with `WHERE attempts > 0` just as effectively, so I'm going with `attempts`.)