perf(worker): avoid sequential scan during worker shutdown
As I've been testing this series of query/index changes, I noticed that things would hang everytime a new deploy went out. This is because when workers shut down, they perform a cleanup step to clear the locked state for any jobs that they had previously locked (and failed to complete, for whatever reason, including a thread crash or timeout).

This `where(locked_by: worker_name)` query was very expensive (resulting in a full sequential scan) as there is no index that targets this column. By scoping the update to the `claimed_by(worker)` scope, we can turn it into an index scan against only live rows.