perf: Improve selectivity of pickup query / "claimed" states
This adjusts pickup queries so that 'run_at <= ?' can be
included as an `Index Cond` (rather than as a `Filter`).

It also improves a few monitor queries in minor ways that
largely improve the `cost` (not visible in our snapshots,
but observed by me in a live testing context).

The core conceit is that `claimable` and `claimed` now
inherently assume `run_at < NOW`. (If a job is scheduled
for the future, it will neither be considered `claimed`
nor `claimable`.)

Note that I also:
- Renamed `unclaimed` to `claimable`
- Removed `workable` as it is synonymous with `claimable`.
- Renamed `workable_by` to `claimable_by`

I did **not** adjust any `Delayed::Monitor` metric names.
Several of these are already outdated and will merit a
migration path, but for now I'm not addressing them.