perf: Improve selectivity of pickup query / "claimed" states (#80)
This adjusts pickup queries so that `run_at <= ?` can be included as an `Index Cond` (rather than as a `Filter`).

It also improves a few monitor queries in minor ways that signifcantly improve the `cost` (not visible in our snapshots, but observed by me in a live testing context).

The core conceit is that `claimable` and `claimed` now inherently assume `run_at < NOW`. (As in: if a job is scheduled for the future, it can neither be considered `claimed` nor `claimable` **even if the job has already been marked as locked by a given worker**.)

By expressing this constraint in the query logic, we enable `run_at < NOW` to supersede anything related to `locked_at`/`locked_by`, enabling the query planner to make more direct use of the index on `run_at` to select fewer rows before applying any in-memory filtering.

(In practice, very little has changed, because the only way to get in that "future claimed" state would be for some manual action to have been applied to `run_at` after a job was locked by a worker.)

Note that I also:
- Renamed `unclaimed` to `claimable`
- Removed `workable` as it is synonymous with `claimable`.
- Renamed `workable_by` to `claimable_by`

I did **not** adjust any `Delayed::Monitor` metric names. Several of these are already outdated and will merit a migration path, but for now I'm not addressing them.