fix: prevent DST + `:local` from breaking job backoff
When using `ActiveRecord.default_timezone = :local` and a local timezone
that has a daylight savings "fall back" hour, it was possible for
future-scheduled jobs to be immediately claimable.

This results from the fact that the DB schema's timestamp column does
not include time zone information, so when Rails encodes a "standard
time" timestamp while `Time.zone.now` is still a DST timestamp (e.g.
`1:59AM EDT`), the future-scheduled `run_at` (e.g. `1:01AM EST`) will
get picked up immediately (since `1:59AM > 1:01AM`).

Naturally, there are a couple ways to avoid this:
- Use `ActiveRecord.default_timezone = :utc` (recommended)
- Use `timestamptz` columns where supported

But if, for legacy reasons, a queue happens to be running in the exact
wrong configuration, the result will be that **future-scheduled jobs
will be deferred to immediately _after_ the "fall back" hour.**

(Jobs that do not specify a `run_at` are not affected and will still be
claimable immediately.)