fix: Improve `upsert_index` to only drop+rebuild if there is no matching & valid index (#83)
This allows for proactively adding indexes out of band, so that the migrator will no-op and complete quickly. For very large tables this may be necessary to stay within constraints of synchronous deployment steps (e.g. a timeout on migration processes).

I also adjusted test logger setup so that it relies entirely on log levels, and will log out migration warnings even if debug logging is turned off for the rest of the test suite.