test(migration path): ensure that new queries are compatible with old index (#84)
This PR aims to provide reasonable unit-level assurance that new code can be rolled out _before_ new indexes are applied.

It applies the same snapshotting strategy to all notable query plans, but generates the plans against a schema with the legacy index (and none of the new indexes).

While as-is this is a wall of only added lines, I split it into two commits and manually inserted the query plans from the prior queries (pulled from the `main` branch snapshots prior to any structural query changes).

So for this first comparison, we can look at [this commit](https://github.com/Betterment/delayed/commit/65ec40209a6490fcd8a530254ef84731f0d68d01) to see that all query plans are at least structurally the same (if not better). **This does not account for `cost` changes,** but should give us a good signal to ensure that we don't regress an `Index Scan` into a `Seq Scan` (for example).